Saving purchase in table

+++ Ticket_Agency/Common/DTO/User/User.cs
@@ -19,7 +19,7 @@ namespace Common.DTO
         [DataMember]
         public long AccountNumber { get; set; }
         [DataMember]
-        public List<int> PurchaseHistory { get; set; }
+        public List<long> PurchaseHistory { get; set; }
 
         public User() { }
 
@@ -38,7 +38,7 @@ namespace Common.DTO
             this.Email = user.Email;
             this.Password = user.Password;
             this.AccountNumber = user.AccountNumber;
-            this.PurchaseHistory = new List<int>();
+            this.PurchaseHistory = new List<long>();
         }
     }
 }
